Add reset encodings and remove findings buttons
