+++ src/hope_dedup_engine/apps/api/admin/__init__.py
@@ -2,5 +2,6 @@ from .deduplicationset import DeduplicationSetAdmin  # noqa
 from .finding import FindingAdmin  # noqa
 from .hdetoken import HDETokenAdmin  # noqa
 from .ignored_pair import IgnoredFilenamePairAdmin, IgnoredReferencePkPairAdmin  # noqa
-from .image import EncodingAdmin  # noqa
+from .encoding import EncodingAdmin  # noqa
 from .jobs import DedupJob  # noqa
+from .deduplicationsetgroup import DeduplicationSetGroupAdmin  # noqa
